Refactoring the few shot management (#272)
---------
