+++ src/lighteval/tasks/lighteval_task.py
@@ -28,11 +28,9 @@ from multiprocessing import Pool
 from pathlib import Path
 from typing import TYPE_CHECKING, Callable, Dict, List, Optional, Tuple, Union
 
-from datasets import load_dataset
 from huggingface_hub import TextGenerationInputGrammarType
 from pytablewriter import MarkdownTableWriter
 
-from lighteval.few_shot_manager import FewShotSampler
 from lighteval.logging.hierarchical_logger import hlog, hlog_warn
 from lighteval.metrics import (
     apply_generative_metric,
@@ -44,6 +42,7 @@ from lighteval.metrics import (
 )
 from lighteval.metrics.metrics import Metric, MetricCategory, Metrics
 from lighteval.models.base_model import BaseModel
+from lighteval.tasks.prompt_manager import PromptManager
 from lighteval.tasks.requests import (
     Doc,
     GreedyUntilMultiTurnRequest,
@@ -55,7 +54,7 @@ from lighteval.tasks.requests import (
     RequestType,
     SampleUid,
 )
-from lighteval.utils.utils import as_list
+from lighteval.utils.utils import as_list, download_dataset_worker
 
 
 if TYPE_CHECKING:
@@ -134,10 +133,35 @@ class LightevalTaskConfig:
         self.suite = tuple(self.suite) if self.suite is not None else None
         self.stop_sequence = tuple(self.stop_sequence) if self.stop_sequence is not None else None
 
+    def print(self):
+        md_writer = MarkdownTableWriter()
+        md_writer.headers = ["Key", "Value"]
+
+        values = []
+
+        for k, v in asdict(self).items():
+            if k == "metric":
+                for ix, metrics in enumerate(v):
+                    for metric_k, metric_v in metrics.items():
+                        if inspect.ismethod(metric_v):
+                            values.append([f"{k} {ix}: {metric_k}", metric_v.__qualname__])
+                        else:
+                            values.append([f"{k} {ix}: {metric_k}", repr(metric_v)])
+
+            else:
+                if isinstance(v, Callable):
+                    values.append([k, v.__name__])
+                else:
+                    values.append([k, repr(v)])
+
+        md_writer.value_matrix = values
+
+        print(md_writer.dumps())
+
 
 class LightevalTask:
     def __init__(  # noqa: C901
-        self, name: str, cfg: LightevalTaskConfig, cache_dir: Optional[str] = None, custom_tasks_module: list = None
+        self, name: str, cfg: LightevalTaskConfig, cache_dir: Optional[str] = None
     ):
         """
         Initialize a LightEval task.
@@ -148,8 +172,6 @@ class LightevalTask:
                 task-specific settings (from the task_table.json file).
             cache_dir (Optional[str], optional): directory to cache the
                 dataset. Defaults to None.
-            custom_tasks_module ([type], optional): A custom module
-                containing task-specific functions. Defaults to None.
         """
         self.name = name
         self.version = cfg.version
@@ -178,9 +200,7 @@ class LightevalTask:
             self.fewshot_split = as_list(cfg.few_shots_split)
         else:
             self.fewshot_split = as_list(self.get_first_possible_fewshot_splits())
-        self.fewshot_sampler = FewShotSampler(
-            few_shots_select=cfg.few_shots_select, few_shots_split=self.fewshot_split
-        )
+        self.fewshot_selection = cfg.few_shots_select
 
         # Metrics
         self.metrics = as_list(cfg.metric)
@@ -225,65 +245,6 @@ class LightevalTask:
     @property
     def cfg(self):
         return self._cfg
-
-    def print_config(self):
-        md_writer = MarkdownTableWriter()
-        md_writer.headers = ["Key", "Value"]
-
-        values = []
-
-        for k, v in asdict(self.cfg).items():
-            if k == "metric":
-                for ix, metrics in enumerate(v):
-                    for metric_k, metric_v in metrics.items():
-                        if inspect.ismethod(metric_v):
-                            values.append([f"{k} {ix}: {metric_k}", metric_v.__qualname__])
-                        else:
-                            values.append([f"{k} {ix}: {metric_k}", repr(metric_v)])
-
-            else:
-                if isinstance(v, Callable):
-                    values.append([k, v.__name__])
-                else:
-                    values.append([k, repr(v)])
-            # print(k, ":", repr(v))
-
-        md_writer.value_matrix = values
-
-        print(md_writer.dumps())
-
-    def doc_to_text_without_instructions(self, doc: Doc) -> str:
-        """
-        Returns the query of the document without the instructions. If the
-        document has instructions, it removes them from the query:
-
-        Args:
-            doc (Doc): document class, containing the query and the
-                instructions.
-
-        Returns:
-            str: Query of the document without the instructions.
-        """
-        if not doc.query.startswith(doc.instruction):
-            raise ValueError(f"Prompt query {doc.query} is not starting with instruction {doc.instruction}")
-        return doc.query[len(doc.instruction) :]
-
-    def doc_to_text_and_instructions(self, doc: Doc) -> Tuple[str, str]:
-        """
-        Returns a tuple with the query of the document and the instructions.
-        If the document has no instructions, the second element of the tuple is
-        an empty string.
-
-        Args:
-            doc (Doc): document, containing the query and the instructions.
-
-        Returns:
-            Tuple[str, str]: A tuple with the query of the document and the
-                instructions.
-        """
-        if not doc.query.startswith(doc.instruction):
-            raise ValueError(f"Prompt query {doc.query} is not starting with instruction {doc.instruction}")
-        return (doc.query[len(doc.instruction) :], doc.instruction)
 
     def get_first_possible_fewshot_splits(self, number_of_splits: int = 1) -> list[str]:
         """
@@ -383,21 +344,6 @@ class LightevalTask:
             if self.must_remove_duplicate_docs:
                 self._docs = self.remove_duplicate_docs(self._docs)
         return self._docs
-
-    def doc_to_target(self, formatted_doc: Doc, few_shot: bool = False) -> str:
-        """
-        Returns the target of the given document.
-
-        Args:
-            formatted_doc (Doc): Formatted document.
-            few_shot (bool, optional): Whether the document is used for few
-                shot examples. Defaults to False.
-
-        Returns:
-            str: Target of the document, which is the correct answer for a document.
-        """
-        # likely we mostly need one example not all
-        return as_list(formatted_doc.get_golds(few_shot=few_shot))[0]
 
     # Requests
     def get_request_type(self) -> list[RequestType]:  # noqa C901
@@ -618,23 +564,6 @@ class LightevalTask:
             task.dataset = dataset
 
 
-def download_dataset_worker(args):
-    """
-    Worker function to download a dataset from the HuggingFace Hub.
-    Used for parallel dataset loading.
-    """
-    dataset_path, dataset_config_name, trust_dataset = args
-    dataset = load_dataset(
-        path=dataset_path,
-        name=dataset_config_name,
-        data_dir=None,
-        cache_dir=None,
-        download_mode=None,
-        trust_remote_code=trust_dataset,
-    )
-    return dataset
-
-
 def create_requests_from_tasks(  # noqa: C901
     task_dict: dict[str, LightevalTask],
     fewshot_dict: dict[str, list[Tuple[int, bool]]],
@@ -692,45 +621,29 @@ def create_requests_from_tasks(  # noqa: C901
         rnd.seed(42)
         rnd.shuffle(task_docs)
 
-        seeds = task.fewshot_sampler.get_fewshot_seeds(num_fewshot_seeds)
+        prompt_manager = PromptManager(lm=lm, task=task)
+        seeds = prompt_manager.few_shot_sampler.get_fewshot_seeds(num_fewshot_seeds)
 
         # We can do several round of fewshots sampling to get some variance informations
         for seed in seeds:
             for doc_id in range(n_samples):
                 doc_id_seed = f"{doc_id}_{seed}"  # if we do several rounds of few shot sampling we have several seeds
                 for num_fewshot, truncate_few_shots in fewshot_dict[task_name]:
-                    # @clefourrier this mechanism does not work if we have the same task n times with different few shot numbers
-                    # to fix!!
-                    cur_task_name = f"{task_name}|{num_fewshot}"
                     doc = task_docs[doc_id]
-                    is_multi_turn = doc.specific is not None and len(doc.specific.get("multi_turn_queries", [])) > 0
-
-                    if is_multi_turn:
-                        ctx, num_effective_few_shots = task.fewshot_sampler.create_multi_turn_contexts(
-                            doc, use_chat_template, system_prompt, lm.tokenizer
-                        )
-                        doc.specific["multi_turn_queries_context"] = ctx
-                    else:
-                        ctx, num_effective_few_shots = task.fewshot_sampler.fewshot_context(
-                            task=task,
-                            doc=doc,
-                            num_fewshot=num_fewshot,
-                            seed=seed,
-                            truncate_few_shots=truncate_few_shots,
-                            max_model_length=lm.max_length,
-                            sampler=rnd,
-                            tokenizer=lm.tokenizer,
-                            use_chat_template=use_chat_template,
-                            system_prompt=system_prompt,
-                        )
-
-                    doc.num_effective_few_shots = num_effective_few_shots
-                    doc.num_asked_few_shots = num_fewshot
-                    doc.ctx = ctx
+                    doc = prompt_manager.add_context_to_doc(
+                        doc,
+                        num_fewshot=num_fewshot,
+                        seed=seed,
+                        sampler=rnd,
+                        truncate_few_shots=truncate_few_shots,
+                        use_chat_template=use_chat_template,
+                        system_prompt=system_prompt,
+                    )
 
                     # Constructing the requests
+                    cur_task_name = f"{task_name}|{num_fewshot}"
                     docs[SampleUid(cur_task_name, doc_id_seed)] = doc
-                    reqs: dict = task.construct_requests(doc, ctx, doc_id_seed, cur_task_name)
+                    reqs: dict = task.construct_requests(doc, doc.ctx, doc_id_seed, cur_task_name)
                     for req_type in req_types:
                         requests[req_type].extend(reqs[req_type])
